refactor async utils
