+++ pyrefly.toml
@@ -40,6 +40,9 @@ project-includes = [
   "examples/custom_parallel/llama_nemotron_super_49b_custom_plan.py",
   "nemo_rl/algorithms/__init__.py",
   "nemo_rl/algorithms/advantage_estimator.py",
+  "nemo_rl/algorithms/async_utils/__init__.py",
+  "nemo_rl/algorithms/async_utils/interfaces.py",
+  "nemo_rl/algorithms/async_utils/replay_buffer.py",
   "nemo_rl/algorithms/logits_sampling_utils.py",
   "nemo_rl/algorithms/loss/__init__.py",
   "nemo_rl/algorithms/loss/interfaces.py",
@@ -107,6 +110,15 @@ project-includes = [
   "nemo_rl/evals/answer_parsing.py",
   "nemo_rl/experience/__init__.py",
   "nemo_rl/experience/rollouts.py",
+  "nemo_rl/modelopt/__init__.py",
+  "nemo_rl/modelopt/models/__init__.py",
+  "nemo_rl/modelopt/models/generation/__init__.py",
+  "nemo_rl/modelopt/models/generation/vllm_quant_worker.py",
+  "nemo_rl/modelopt/models/policy/__init__.py",
+  "nemo_rl/modelopt/models/policy/workers/__init__.py",
+  "nemo_rl/modelopt/models/policy/workers/dtensor_quant_policy_worker.py",
+  "nemo_rl/modelopt/models/policy/workers/dtensor_quant_policy_worker_v2.py",
+  "nemo_rl/modelopt/registry.py",
   "nemo_rl/models/__init__.py",
   "nemo_rl/models/automodel/__init__.py",
   "nemo_rl/models/dtensor/__init__.py",
@@ -127,15 +139,6 @@ project-includes = [
   "nemo_rl/models/policy/utils.py",
   "nemo_rl/models/policy/workers/__init__.py",
   "nemo_rl/models/policy/workers/patches.py",
-  "nemo_rl/modelopt/__init__.py",
-  "nemo_rl/modelopt/models/__init__.py",
-  "nemo_rl/modelopt/models/generation/__init__.py",
-  "nemo_rl/modelopt/models/generation/vllm_quant_worker.py",
-  "nemo_rl/modelopt/models/policy/__init__.py",
-  "nemo_rl/modelopt/models/policy/workers/__init__.py",
-  "nemo_rl/modelopt/models/policy/workers/dtensor_quant_policy_worker.py",
-  "nemo_rl/modelopt/models/policy/workers/dtensor_quant_policy_worker_v2.py",
-  "nemo_rl/modelopt/registry.py",
   "nemo_rl/utils/__init__.py",
   "nemo_rl/utils/checkpoint.py",
   "nemo_rl/utils/config.py",
