+++ tests/unit/algorithms/test_async_utils.py
@@ -113,16 +113,12 @@ class TestReplayBuffer:
 
         # Push trajectories
         status1 = ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory1, weight_version=0, target_weight_version=1
-            )
+            buffer.add.remote(trajectory1, weight_version=0, target_weight_version=1)
         )
         assert status1 == "success"
 
         status2 = ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory2, weight_version=1, target_weight_version=2
-            )
+            buffer.add.remote(trajectory2, weight_version=1, target_weight_version=2)
         )
         assert status2 == "success"
 
@@ -149,23 +145,17 @@ class TestReplayBuffer:
 
         # Push first two trajectories
         status1 = ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory1, weight_version=0, target_weight_version=1
-            )
+            buffer.add.remote(trajectory1, weight_version=0, target_weight_version=1)
         )
         status2 = ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory2, weight_version=1, target_weight_version=2
-            )
+            buffer.add.remote(trajectory2, weight_version=1, target_weight_version=2)
         )
         assert status1 == "success"
         assert status2 == "success"
 
         # Try to push third trajectory (should return "full")
         status3 = ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory3, weight_version=2, target_weight_version=3
-            )
+            buffer.add.remote(trajectory3, weight_version=2, target_weight_version=3)
         )
         assert status3 == "full"
 
@@ -188,7 +178,7 @@ class TestReplayBuffer:
             }
             trajectories.append(trajectory)
             ray.get(
-                buffer.push_with_wait_signal.remote(
+                buffer.add.remote(
                     trajectory, weight_version=i, target_weight_version=i + 1
                 )
             )
@@ -220,9 +210,7 @@ class TestReplayBuffer:
         # Push only one trajectory
         trajectory = {"batch": {"data": "test"}, "rollout_metrics": {"reward": 1.0}}
         ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory, weight_version=0, target_weight_version=1
-            )
+            buffer.add.remote(trajectory, weight_version=0, target_weight_version=1)
         )
 
         # Try to sample more trajectories than available for current step
@@ -250,12 +238,10 @@ class TestReplayBuffer:
         }
 
         ray.get(
-            buffer.push_with_wait_signal.remote(
-                old_trajectory, weight_version=0, target_weight_version=1
-            )
+            buffer.add.remote(old_trajectory, weight_version=0, target_weight_version=1)
         )
         ray.get(
-            buffer.push_with_wait_signal.remote(
+            buffer.add.remote(
                 recent_trajectory, weight_version=2, target_weight_version=3
             )
         )
@@ -290,14 +276,10 @@ class TestReplayBuffer:
         }
 
         ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory1, weight_version=0, target_weight_version=1
-            )
+            buffer.add.remote(trajectory1, weight_version=0, target_weight_version=1)
         )
         ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory2, weight_version=1, target_weight_version=2
-            )
+            buffer.add.remote(trajectory2, weight_version=1, target_weight_version=2)
         )
 
         # Sample for current step 1 - should only get trajectory intended for step 1
@@ -328,14 +310,10 @@ class TestReplayBuffer:
         trajectory2 = {"batch": {"data": "test2"}, "rollout_metrics": {"reward": 2.0}}
 
         ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory1, weight_version=0, target_weight_version=1
-            )
+            buffer.add.remote(trajectory1, weight_version=0, target_weight_version=1)
         )
         ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory2, weight_version=1, target_weight_version=3
-            )
+            buffer.add.remote(trajectory2, weight_version=1, target_weight_version=3)
         )
 
         existing_weights = ray.get(buffer.get_existing_target_weights.remote())
@@ -350,9 +328,7 @@ class TestReplayBuffer:
         # Push some trajectories
         trajectory = {"batch": {"data": "test"}, "rollout_metrics": {"reward": 1.0}}
         ray.get(
-            buffer.push_with_wait_signal.remote(
-                trajectory, weight_version=0, target_weight_version=1
-            )
+            buffer.add.remote(trajectory, weight_version=0, target_weight_version=1)
         )
 
         # Verify buffer has content
@@ -645,7 +621,7 @@ class TestAsyncUtilsIntegration:
                 "rollout_metrics": {"reward": float(trajectory_id)},
             }
             return ray.get(
-                buffer.push_with_wait_signal.remote(
+                buffer.add.remote(
                     trajectory,
                     weight_version=trajectory_id,
                     target_weight_version=trajectory_id + 1,
